+++ Model/EntidadesProducto/Relleno.cs
@@ -4,13 +4,13 @@
 	{
 		public Relleno() { }
 
-		public Relleno(string name, double precio)
+		public Relleno(string name, decimal precio)
 		{
 			Nombre = name;
 			Precio = precio;
 		}
 
-		public Relleno(string name, double precio, int id)
+		public Relleno(string name, decimal precio, int id)
 		{
 			Nombre = name;
 			Precio = precio;
